// server/generated/model/message.ts
/**
 * Chat Sage API
 * Slack-like messaging platform with universal soft-deletes, selective partitioning, email-based verification, triggers for updatedAt, real-time features, error code enumeration, and recommended monorepo structure.
 *
 * The version of the OpenAPI document: v1.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class Message {
    'messageId'?: number;
    'userId'?: number | null;
    'channelId'?: number | null;
    'workspaceId'?: number | null;
    'parentMessageId'?: number | null;
    'content'?: string;
    'deleted'?: boolean;
    'postedAt'?: Date;
    'createdAt'?: Date;
    'updatedAt'?: Date;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "messageId",
            "baseName": "messageId",
            "type": "number"
        },
        {
            "name": "userId",
            "baseName": "userId",
            "type": "number"
        },
        {
            "name": "channelId",
            "baseName": "channelId",
            "type": "number"
        },
        {
            "name": "workspaceId",
            "baseName": "workspaceId",
            "type": "number"
        },
        {
            "name": "parentMessageId",
            "baseName": "parentMessageId",
            "type": "number"
        },
        {
            "name": "content",
            "baseName": "content",
            "type": "string"
        },
        {
            "name": "deleted",
            "baseName": "deleted",
            "type": "boolean"
        },
        {
            "name": "postedAt",
            "baseName": "postedAt",
            "type": "Date"
        },
        {
            "name": "createdAt",
            "baseName": "createdAt",
            "type": "Date"
        },
        {
            "name": "updatedAt",
            "baseName": "updatedAt",
            "type": "Date"
        }    ];

    static getAttributeTypeMap() {
        return Message.attributeTypeMap;
    }
}